Move ConsoleHelpers to its own source file
